fix: Correct type mismatches across multiple modules
Fix all remaining build errors in the codebase:

**incidentresponse/keeper/invariants.go:**
- Cast float64 to int64 for fmt.Sprintf %d format

**confidencescore/types/conversions_test.go:**
- Replace timestamppb.Timestamp with gogoproto types.Timestamp
- Use types.TimestampProto() to create timestamps

**cryptography/keeper:**
- Replace timestamppb with time.Time for stdtime fields
- Fix Params type from pointer to value in GenesisState
- Update all test files to use correct timestamp types

**dataregistry/keeper/query_server_test.go:**
- Fix timestamp type usage
- Correct type alias references from types to drtypes

**economics/types/codec_test.go:**
- Simplify test to work with standard protobuf types
- Remove ProtoMarshaler interface usage

**dex/types:**
- Fix type comparisons for LegacyDec and Int types
- Use .Equal() method instead of string comparison
- Update params_test.go and security_types_test.go

All modules now build successfully without errors.

+++ chain/x/incidentresponse/keeper/invariants.go
@@ -78,7 +78,7 @@ func ParamsInvariant(k *Keeper) sdk.Invariant {
 					types.ModuleName,
 					"params-valid",
 					fmt.Sprintf("max pause duration too long: %d seconds (max 7 days)",
-						params.MaxPauseDuration.Seconds()),
+						int64(params.MaxPauseDuration.Seconds())),
 				), true
 			}
 		}
